split: extract CredentialFlowService from UserSelfServiceService
Moves token+email driven credential flows (email verification, forgot
password, invite, forced password change, magic link, account-locked
notification) into a focused service. UserSelfServiceService keeps only
profile, password change, and session management methods.

AuthService, AdminAccountService, and AdminUserService repoint their
selfServiceService dependency to credentialFlowService since they all
invoke credential flows. PortalRoutes is the only consumer that keeps
the UserSelfServiceService dependency.
