+++ src/main/kotlin/com/kauth/adapter/web/auth/AcceptInviteRoutes.kt
@@ -2,8 +2,8 @@ package com.kauth.adapter.web.auth
 
 import com.kauth.domain.model.SecurityConfig
 import com.kauth.domain.port.RateLimiterPort
+import com.kauth.domain.service.CredentialFlowService
 import com.kauth.domain.service.SelfServiceResult
-import com.kauth.domain.service.UserSelfServiceService
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.html.respondHtml
 import io.ktor.server.request.receiveParameters
@@ -13,7 +13,7 @@ import io.ktor.server.routing.get
 import io.ktor.server.routing.post
 
 internal fun Route.acceptInviteRoutes(
-    selfServiceService: UserSelfServiceService,
+    credentialFlowService: CredentialFlowService,
     rateLimiter: RateLimiterPort,
 ) {
     get("/accept-invite") {
@@ -59,7 +59,7 @@ internal fun Route.acceptInviteRoutes(
             )
         }
 
-        when (val result = selfServiceService.confirmAcceptInvite(token, newPassword, confirmPassword)) {
+        when (val result = credentialFlowService.confirmAcceptInvite(token, newPassword, confirmPassword)) {
             is SelfServiceResult.Success ->
                 call.respondHtml(
                     HttpStatusCode.OK,
