+++ src/test/kotlin/com/kauth/adapter/web/api/ApiRoutesTest.kt
@@ -15,8 +15,8 @@ import com.kauth.domain.model.UserId
 import com.kauth.domain.service.AdminAccountService
 import com.kauth.domain.service.AdminResult
 import com.kauth.domain.service.ApiKeyService
+import com.kauth.domain.service.CredentialFlowService
 import com.kauth.domain.service.RoleGroupService
-import com.kauth.domain.service.UserSelfServiceService
 import com.kauth.fakes.FakeApiKeyRepository
 import com.kauth.fakes.FakeApplicationRepository
 import com.kauth.fakes.FakeAuditLogPort
@@ -121,7 +121,7 @@ class ApiRoutesTest {
         )
 
     private val selfServiceService =
-        UserSelfServiceService(
+        CredentialFlowService(
             userRepository = userRepo,
             tenantRepository = tenantRepo,
             sessionRepository = sessionRepo,
@@ -138,7 +138,7 @@ class ApiRoutesTest {
             tenantRepository = tenantRepo,
             userRepository = userRepo,
             auditLog = auditLogPort,
-            selfServiceService = selfServiceService,
+            credentialFlowService = selfServiceService,
         )
 
     private val adminUserService =
@@ -148,7 +148,7 @@ class ApiRoutesTest {
             sessionRepository = sessionRepo,
             passwordHasher = hasher,
             auditLog = auditLogPort,
-            selfServiceService = selfServiceService,
+            credentialFlowService = selfServiceService,
         )
 
     private val applicationManagementService =
