+++ src/test/kotlin/com/kauth/adapter/web/auth/EmailOtpLoginRoutesTest.kt
@@ -10,9 +10,9 @@ import com.kauth.domain.model.TenantTheme
 import com.kauth.domain.model.User
 import com.kauth.domain.model.UserId
 import com.kauth.domain.service.AuthService
+import com.kauth.domain.service.CredentialFlowService
 import com.kauth.domain.service.EmailOtpService
 import com.kauth.domain.service.OAuthService
-import com.kauth.domain.service.UserSelfServiceService
 import com.kauth.fakes.FakeApplicationRepository
 import com.kauth.fakes.FakeAuditLogPort
 import com.kauth.fakes.FakeAuthorizationCodeRepository
@@ -113,7 +113,7 @@ class EmailOtpLoginRoutesTest {
         )
 
     private fun selfService() =
-        UserSelfServiceService(
+        CredentialFlowService(
             userRepository = users,
             tenantRepository = tenants,
             sessionRepository = sessions,
@@ -179,7 +179,7 @@ class EmailOtpLoginRoutesTest {
                     loginRateLimiter = limiter,
                     registerRateLimiter = limiter,
                     tokenRateLimiter = limiter,
-                    selfServiceService = selfService(),
+                    credentialFlowService = selfService(),
                     encryptionService = encryptionService,
                     translationPort = EnglishOnlyTranslation(),
                     emailOtpService = otpService(),
