+++ src/test/kotlin/com/kauth/adapter/web/auth/SilentAuthTest.kt
@@ -10,9 +10,9 @@ import com.kauth.domain.model.TenantTheme
 import com.kauth.domain.model.User
 import com.kauth.domain.model.UserId
 import com.kauth.domain.service.AuthService
+import com.kauth.domain.service.CredentialFlowService
 import com.kauth.domain.service.MfaService
 import com.kauth.domain.service.OAuthService
-import com.kauth.domain.service.UserSelfServiceService
 import com.kauth.fakes.FakeApplicationRepository
 import com.kauth.fakes.FakeAuditLogPort
 import com.kauth.fakes.FakeAuthorizationCodeRepository
@@ -59,7 +59,7 @@ class SilentAuthTest {
     private val tokenPort = FakeTokenPort()
     private val limiter = InMemoryRateLimiter(maxRequests = 1000, windowSeconds = 60)
     private val mfaService = mockk<MfaService>(relaxed = true)
-    private val selfService = mockk<UserSelfServiceService>(relaxed = true)
+    private val selfService = mockk<CredentialFlowService>(relaxed = true)
     private val encryptionService = EncryptionService("test-secret-key-32-chars-minimum-len")
 
     private val tenant =
@@ -143,7 +143,7 @@ class SilentAuthTest {
                     loginRateLimiter = limiter,
                     registerRateLimiter = limiter,
                     tokenRateLimiter = limiter,
-                    selfServiceService = selfService,
+                    credentialFlowService = selfService,
                     mfaService = mfaService,
                     encryptionService = encryptionService,
                     translationPort = EnglishOnlyTranslation(),
